requirements.txt
langchain
langchain-community
langchain_ollama
langchain_core
langgraph
duckduckgo-search
faiss-cpu
pypdf